+++ packages/sandbox/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Minor Changes
 
-- Support new keepLast snapshots parameters for CLI and SDK
+- Support new keepLastSnapshots parameter for CLI and SDK
 
 ### Patch Changes